+++ src/boot/router.js
@@ -1,8 +1,12 @@
+import { handleHistory } from '../helpers/historyHandler'
+
 export default async ({ Vue, router }) => {
   Vue.prototype.$history = []
+  const history = handleHistory()
 
   router.beforeEach(async (to, from, next) => {
     Vue.prototype.$history.unshift(from)
+    history.push(to)
     next()
   })
 }