+++ src/components/FormView.vue
@@ -23,6 +23,8 @@
       <slot name="footer" :errors="errors" :fields="fields" :metadata="metadata"/>
     </footer>
 
+    <qs-dialog v-model="showDialog" v-bind="dialogConfig" />
+
     <q-inner-loading :showing="isFetching">
       <q-spinner color="grey" size="3em" />
     </q-inner-loading>
@@ -30,8 +32,9 @@
 </template>
 
 <script>
-import { get, isEqual } from 'lodash'
+import { get, isEqual, cloneDeep } from 'lodash'
 import { extend } from 'quasar'
+import { handleHistory } from '../helpers/historyHandler'
 
 import viewMixin from '../mixins/view'
 
@@ -91,7 +94,23 @@ export default {
   data () {
     return {
       isSubmiting: false,
-      cachedResult: {}
+      cachedResult: {},
+      hasResult: false,
+      showDialog: false,
+      dialogConfig: {
+        card: {
+          title: 'Atenção!',
+          description: 'Você está deixando a página e suas alterações serão perdidas. Tem certeza que deseja sair sem salvar?'
+        },
+        ok: {
+          props: { label: 'Continuar editando' },
+          events: null
+        },
+        cancel: {
+          props: { label: 'Sair' },
+          events: null
+        }
+      }
     }
   },
 
@@ -135,8 +154,12 @@ export default {
 
   watch: {
     fields (fields) {
-      const models = this.getModelsByFields(fields)
-      this.$emit('input', { ...models, ...this.value })
+      const models = { ...this.getModelsByFields(fields), ...this.value }
+      this.$emit('input', models)
+
+      if (!this.hasResult && this.showDialogOnUnsavedChanges) {
+        this.cachedResult = cloneDeep(models)
+      }
     }
   },
 
@@ -166,8 +189,9 @@ export default {
         this.setMetadata(metadata)
 
         if (result) {
-          this.cachedResult = this.showDialogOnUnsavedChanges && extend(true, {}, result)
+          this.hasResult = true
           this.$emit('input', { ...this.value, ...result })
+          this.cachedResult = this.showDialogOnUnsavedChanges && extend(true, {}, result)
         }
 
         this.$emit('fetch-success', response, this.value)
@@ -206,6 +230,10 @@ export default {
           { id: this.id, payload: this.value, url: this.url }
         )
 
+        if (this.showDialogOnUnsavedChanges) {
+          this.cachedResult = cloneDeep(this.value)
+        }
+
         this.setErrors()
         this.$qs.success(response.data.status.text || 'Item salvo com sucesso!')
         this.$emit('submit-success', response, this.value)
@@ -226,6 +254,17 @@ export default {
       }
     },
 
+    openDialog () {
+      this.showDialog = true
+    },
+
+    handleDialog (fn) {
+      this.openDialog()
+
+      this.dialogConfig.ok.events = { click: () => handleHistory().push(this.$route) }
+      this.dialogConfig.cancel.events = { click: fn }
+    },
+
     beforeRouteLeave (to, from, next) {
       if (!this.showDialogOnUnsavedChanges) {
         return null
@@ -235,22 +274,7 @@ export default {
         return next()
       }
 
-      this.$q.dialog({
-        title: 'Atenção',
-        message: 'Você está deixando a página e suas alterações serão perdidas. Tem certeza que deseja sair sem salvar?',
-        persistent: true,
-        cancel: {
-          label: 'Sair',
-          noCaps: true,
-          outline: true
-        },
-        ok: {
-          noCaps: true,
-          label: 'Continuar editando'
-        }
-      }).onCancel(() => {
-        next()
-      }).onOk(() => next(false))
+      this.handleDialog(next)
     },
 
     handleCancelRoute () {